requirements.txt
Django==1.4.1
South==0.7.6
boto==2.1.0
bpython==0.11
coverage==3.5.2
distribute==0.6.27
dj-database-url==0.2.1
django-braces==0.1.3
django-cache-machine==0.6
django-coverage==1.2.2
django-crispy-forms==1.1.4
django-debug-toolbar==0.9.4
django-extensions==0.8
django-floppyforms==1.0
django-heroku-memcacheify==0.3
django-heroku-postgresify==0.2
django-jsonfield==0.8.11
django-profiletools==0.2.1
django-pylibmc-sasl==0.2.4
django-registration==0.8
django-storages==1.1.3
django-wysiwyg==0.4.1
gunicorn==0.14.2
psycopg2
pylibmc==1.2.3
wsgiref==0.1.2
newrelic
raven